Move devnode to t/devtests to avoid cycle in imports
